Create new parser module
Use a __init__.py file to create a new module named parser, also start
working on the wall parser.

The wall parser reads the wall.html file from a Facebook backup and
parses through it aggregating the data out of the HTML into another
format, at least that is the plan.

The main parser structure/statemachine is in place more work has to be
done in an attempt to finish up the different entries and to actually
get it parsing the data it comes across.